fix: `test-e2e-chrome-vault-decryption` job (#31425)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

The Vault Decryptor spec is flaky. The problem is that whenever we
started the spec, we wait until the MetaMask full window was open and
then switch to it, just by switching to the last open window (broken
assumption: last window is the MetaMask full width one). The reason for
not using the regular driver methods, (switch by title) was because we
there's no socket for windows in the prod build.
This behavior was stable, but after this [PR was
merged](https://github.com/MetaMask/metamask-extension/pull/30970) this
changed.

Now the window handles we get in the array are sometimes unordered (the
MetaMask full width is not the 3rd one but sometimes the 2nd, and the
offscreen is the 3rd -- see video below).

The fix is to make the switch to MetaMask deterministic, by switching to
each window handle and getting the window title and check if it matches
the full width one.


https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/134516/workflows/ec495927-6baf-4324-86ff-b65786197dad/jobs/4703080/steps

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/31425?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. successful ci run here:
https://app.circleci.com/pipelines/github/MetaMask/metamask-extension/134537/workflows/7095dd73-ff37-4a6f-b35a-b3e0e7fe9bc6/jobs/4703479

## **Screenshots/Recordings**

You can see how the error says it doesn't find the title, but in the
artifacts we do see it, meaning we are not focused on the correct window

![Screenshot from 2025-03-31
09-01-27](https://github.com/user-attachments/assets/1d996911-76df-47fa-a2db-7488ebadb8d0)

See the order of the windows on a failed test:
1. Dom1: chrome Extension page
2. Dom2: MetaMask full width
3. Dom3: MetaMAsk offscreen

And in our spec we assumed the MetaMask full witdth was always at the
last position of the array,



https://github.com/user-attachments/assets/67180246-a464-4d1f-8c9e-6fa391f7736d



## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.