fix: flaky test `MetaMask onboarding should not prevent network requests to basic functionality endpoints when the basic functionality toggle is on` (#31521)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
We have the following duplicated keys in the onboarding fixtures,
causing the wallet to throw errors if we leave it enough time after
onboarding:

```
        showTestNetworks: false,
        smartTransactionsOptInStatus: true,
        tokenNetworkFilter: {},
```

This properties are already inside the `PreferencesController >
preferences` and then they are duplicated in the
`PreferencesController`, which are wrong.



Extra note: the solution to the root cause of this mistake of the
fixtures data schema, should be to add a mechanism to enforce the
fixtures data integrity, and also to use the latest data schema -- even
to support setting dynamically the version we want to test. However this
is out of scope of this PR, but we already have a task for doing so in
the following months.,

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/31521?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. Check ci jobs

(I've added the `only` keyword on the failing test. This way this file
will be rerun x5 times, as well as added the ff and webpack runs)

## **Screenshots/Recordings**

![Screenshot from 2025-04-02
16-53-25](https://github.com/user-attachments/assets/a4174771-b486-4c99-b10f-ae11e0ba3547)


<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.