feat: show error to user if vault should exist, but is missing (#31404)
## **Description**

This PR updates background.js, and some related files, to ensure that an
error is properly shown to the user in two cases:
1) If the vault is missing from state when it should be present
2) For the case of a known corrupted state error with the message "block
checksum mismatch"

### How?

To achieve this, this PR adds logic to detect if the vault is missing
and then throws an error, which in turn is sent via `postmessage` to the
UI, where it is handled by a custom error screen.

It detects a missing vault by first storing a backup copy of it in
`IndexedDB`, later, when reading back out, if the primary database does
NOT have the `vault` it then checks if there is a backup. If there _is_
a backup, it throws the missing vault error.

This PR does _not_ give the user the option to restore the backup vault.
That will come in a follow up PR.

note: the follow up PR that enables vault recovery will target _this_
PR's branch, and will be merged into it after it is reviewed separately.

### What this does NOT do that you might think it can/should do:

We cannot reliably detect if the background.js code is running because
the user just installed MetaMask.

While there is an `onInstalled` event we can listen to, there are no
events telling us it wasn't just installed. So we have no way of waiting
for `onInstalled` to resolve, because it never will in the case that we
weren't installed just then.

But even if we could detect if it was just installed, users don't always
onboard immediately after installing. If they onboard at some other
time, so the `onInstalled` signal is actually not something we can use.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/31404?quickstart=1)

## **Related issues**

Possibly related to:

https://github.com/MetaMask/metamask-extension/issues/23165
https://github.com/MetaMask/metamask-extension/issues/22589
https://github.com/MetaMask/metamask-extension/issues/22533

## **Manual testing steps**

1. Install and onboard
2. Open up the service worker console and enter this code:
```javascript
chrome.storage.local.get(({ data, meta }) => {
  chrome.storage.local.set(
    { data: {} , meta },
    () => { chrome.runtime.reload() }
  );
});
```
3. After opening the UI again, you should see an error message
indicating the state is corrupt


![image](https://github.com/user-attachments/assets/3fed0e10-9af8-4898-b104-4e2d865eb236)


Alternatively
1. Install and onboard
2. Edit one of the level db files in which the extensions data is saved
by deleting a single character in the file
3. Reload the extension
4. After opening the UI again, you should see an error message
indicating the state is corrupt

## **Screenshots/Recordings**



## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.

---------

Co-authored-by: Dan J Miller <danjm.com@gmail.com>
Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>
Co-authored-by: Danica Shen <zhaodanica@gmail.com>