feat: Remove failed transaction retry icon button  (#32024)
## **Description**

There is an issue where after a transaction fails, using the retry
button may result in a successful transaction, but the transaction would
continue to say "Failed" in the transaction list. This seems to be an
outstanding bug.

We are hiding the retry button until further notice. I chose to comment
these out rather than setting showRetry to make it more explicit that it
is unused code.

This retry button was shown in the TransactionListItemDetails when the
failed transaction was neither a swap nor a smart transaction.

The speedup and cancel buttons continue to exist. Speedup reuses the
same logic as the Retry logic.

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/28615

## **Manual testing steps**

1. Submit a failed transaction
  a. e.g. Use Arbitrum One network
  b. Go to the test-dapp
c. Send a EIP-1559 or Legacy Transaction (with gas specified by the
dapp)
2. Go to MM > Activity
3. Click on the Failed transaction

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

![CleanShot 2025-04-11 at 09 40
46](https://github.com/user-attachments/assets/4d1ecb74-2a3a-4674-81ea-c56fcd99f0f2)

### **After**

![CleanShot 2025-04-15 at 18 12
17](https://github.com/user-attachments/assets/b15eee49-4cfb-4bb5-ba06-94e8a2f6b314)

## **Pre-merge author checklist**

- [ ] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [ ] I've completed the PR template to the best of my ability
- [ ] I’ve included tests if applicable
- [ ] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [ ] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.