ci: fix changed-files error (#32633)
## **Description**

If something is on the Merge Queue, or it's a release branch, or it fits
one of `main, master, Version-v*, trigger-ci*`, it will be considered
"not a PR."

In these cases, you will see almost a hundred of these errors in the
GitHub Actions log.


![image](https://github.com/user-attachments/assets/b409935e-3fd4-45fb-b3fa-8f814d9c091a)

It's not actually a real error though, it's expected behavior. So I've
changed the flow to make an artifact of an empty `changed-files.json`
list.

A separate problem is actually do we **WANT** a non-empty
`changed-files.json` in some of these cases, but I'm not going to tackle
that one right now.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/32633?quickstart=1)

<!--## **Related issues**
## **Manual testing steps**
## **Screenshots/Recordings**
## **Pre-merge author checklist**
## **Pre-merge reviewer checklist**-->