ci: better E2E test splitting (#32657)
## **Description**

The goal of this PR is to spend less CPU time and less developer time
(both of which are in short supply) running E2E tests. Look at the
screenshots below, these are totally fair apples-to-apples comparisons
of the runtimes.

We achieve this optimization in two ways:

1. Do not build the `yarn build:test:mv2` 20 times, in each of the
shards, just build it once (this was easy because the prep work was
already done, we just had to finish the work)
2. Tests were previously split across the shards in a completely naive
way. We now use data from the last successful run on `main` to predict
timings and try to achieve an even split. We used to do this on
CircleCI, but that had some flaws, and this method works even better.

- **Q.** Why didn't we use an existing library like
https://github.com/marketplace/actions/split-tests-by-timings or
https://github.com/marketplace/actions/split-tests?
- **A.** We have several tricky elements that none of the libraries
support:
- We have a list of test files, rather than a glob (but if we had to, we
could potentially convert to globs)
- We need to support duplicated test files for the quality gate feature
- The local test files do not always match the reports (because tests
were added or deleted)

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/32657?quickstart=1)

## **Screenshots/Recordings**

### **Before**


![image](https://github.com/user-attachments/assets/dbb5c99c-2ae0-468f-adc7-c86b5d5b0606)

### **After**


![image](https://github.com/user-attachments/assets/2c0e531e-4d72-4443-9ca3-e519ef8dc6e0)

<!--## **Manual testing steps**
## **Related issues**
## **Pre-merge author checklist**
## **Pre-merge reviewer checklist**-->