fix: flaky scroll step in snap tests (#33012)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
In some snap tests, we often see the error `Waiting until element is
enabled` when we want to click Confirm on the Snap installation page.

The problem is that the Confirm button is not enabled, unless we scroll
to the bottom of the page, by clicking the arrow `snap-install-scroll`.
Now, this button is clicked using the method `clickElementSafe` because
depending on the rendering, the arrow appears or not.
Here comes the issue: this method, waits by default 2 seconds, and if
it's not found in the DOM, then we skip to the next step without
failing. Then the next step is clicking Confirm, but the button is not
enabled if that arrow hasn't been clicked. The problem is that whenever
the snap dialog is open, in some tests we directly call the
`clickElementSafe`, but the page can take some time to load. If it takes
more than 2 seconds, then the `clickElementSafe` will skip clicking the
button and then the Confirm is never enabled.
The solution is to first wait for the confirmation page to be loaded, by
looking for a common element, and then trigger the `clickElementSafe`.

Alternatively, we could increase the timeout of `clickElementSafe` but
this is not desirable as this can affect other tests, that don't need
that amount of time, and slow our pipeline.


![test-failure-screenshot-3](https://github.com/user-attachments/assets/3e8cc8cc-00cf-4bc9-93c1-1ab556ccc0c7)

![Screenshot from 2025-05-16
10-29-35](https://github.com/user-attachments/assets/a9d4d93e-a270-404a-9ab6-ac4b9336329c)


[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/33012?quickstart=1)

## **Related issues**

Fixes:

## **Manual testing steps**

1. Go to this page...
2.
3.

## **Screenshots/Recordings**

<!-- If applicable, add screenshots and/or recordings to visualize the
before and after of your change. -->

### **Before**

<!-- [screenshots/recordings] -->

### **After**

<!-- [screenshots/recordings] -->

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.