test: mock snap registry calls by adding the local snap binaries and download script (#32555)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**
A common failure in snap tests is when we try to install any snap, and
we get a bad response from the npm registry server, like this one: `the
registry is temporarily unavailable`

![Screenshot from 2025-05-06
13-55-56](https://github.com/user-attachments/assets/2aa6e3bf-26c9-492a-931a-88b3b5f864d0)


This PR does the following:


- adds all the snap binaries and headers,, so we can return those as a
mocked response instead of relying on 3rd party calls.
- removes the entry from the allow list for the e2e live requests,
meaning that if any snap is updated, this will trigger a new live
request, but it will be catched by the catch-all-mock and will then
avoid a live request and force the person who updated that, to add the
new snap binary for that new version.
- adds a new script for downloading the binary and headers, so the mock
is easier to update, instead of having to do it manually, whenever we
want to update a new version of a snap package in the code.


[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/32555?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/32556

## **Manual testing steps**
Note: for reviewers. The PR is very big but mostly because of the
binaries and headers files. The relevant logic which contains the
changes is in these 2 files:
- `snap-binary-mocks.ts`  -- mock functions, to use in our e2e
- `update-snap-binary.ts` -- script to update a specific snap with a
given name and version

1. All tests should pass


2. For testing the `yarn update-snap-binary` script:
  - run `yarn update-snap-binary --help` for help
- run `yarn update-snap-binary --bip32-example-snap@2.2.0` this will
remove the existing bip32 snap binary and header files and download the
ones specifically for version 2.2.0
- run `yarn update-snap-binary --bip32-example-snap@2.3.0` this will
remove the existing bip32 snap binary and header files and download the
ones specifically for version 2.3.0
- run `yarn update-snap-binary --bip32-example-snap@2.5.0` this will
return an error as this version doesn't exist





https://github.com/user-attachments/assets/bb106eb6-0a40-482a-953e-7c26fbd50789



## **Screenshots/Recordings**
Manual process:


https://github.com/user-attachments/assets/c05716dc-ba93-4a28-aaa8-5f4c6869438d

## **Pre-merge author checklist**

- [x] I've followed [MetaMask Contributor
Docs](https://github.com/MetaMask/contributor-docs) and [MetaMask
Extension Coding
Standards](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/CODING_GUIDELINES.md).
- [x] I've completed the PR template to the best of my ability
- [x] I’ve included tests if applicable
- [x] I’ve documented my code using [JSDoc](https://jsdoc.app/) format
if applicable
- [x] I’ve applied the right labels on the PR (see [labeling
guidelines](https://github.com/MetaMask/metamask-extension/blob/main/.github/guidelines/LABELING_GUIDELINES.md)).
Not required for external contributors.

## **Pre-merge reviewer checklist**

- [ ] I've manually tested the PR (e.g. pull and build branch, run the
app, test code being changed).
- [ ] I confirm that this PR addresses all acceptance criteria described
in the ticket it closes and includes the necessary testing evidence such
as recordings and or screenshots.