fix: problem in `yarn start:with-state` (#34851)
## **Description**

The change to turn backup and sync on by default conflicted with the
`yarn start:with-state` command, because it would automatically sync
with the remote state of the SRP and overwrite what you just did.

This PR turns off backup and sync if you're using `yarn
start:with-state`.

Future plans are to make this work with power user E2E tests.

## **Changelog**

CHANGELOG entry: null

## **Related issues**

Fixes: MetaMask/MetaMask-planning#5451

<!--## **Manual testing steps**
## **Screenshots/Recordings**
## **Pre-merge author checklist**
## **Pre-merge reviewer checklist**-->