perf: Exclude test assets in production (#34690)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Reduces the bundle size of the background bundle by ~9% by removing test
files that aren't required in production.

We are currently including stubs, fixtures and E2E data in the
production builds. The code that brings it in is removed when bundling
for production, but the imports remain due to limited tree-shaking in
Browserify. This PR ensures that these test files are excluded from the
bundle. To do this, I also had to rewrite a couple of the imports.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/34690?quickstart=1)