fix: cp-13.2.0 Make SnapsNameProvider respect the chainIds caveat (#35477)
<!--
Please submit this PR as a draft initially.
Do not mark it as "Ready for review" until the template has been
completely filled out, and PR status checks have passed at least once.
-->

## **Description**

Previously the SnapsNameProvider would not respect the `chainIds` caveat
when specified by the Snap, this would result in calls to the Snap that
it is not expecting and could crash it. In production this crashes the
Solana Snap.

[![Open in GitHub
Codespaces](https://github.com/codespaces/badge.svg)](https://codespaces.new/MetaMask/metamask-extension/pull/35477?quickstart=1)

## **Related issues**

Fixes: https://github.com/MetaMask/metamask-extension/issues/35473
Fixes: https://github.com/MetaMask/metamask-extension/issues/35364